Family Treatment Weight change data for young female anorexia patients.

The MASS package includes the dataset anorexia, containing pre and post treatment weights for young female anorexia patients. This is a subset of those data, containing only those patients who received Family Treatment.

Usage

data(anorexia.sub)

Format

A dataframe with 17 observations on the following 2 variables, no NAs.

Prewt

Pretreatment weight of subject, in pounds.

Postwt

Postreatment weight of subject, in pounds.

Source

Hand, D. J., Daly, F., McConway, K., Lunn, D. and Ostrowski, E. eds (1993) A Handbook of Small Data Sets. Chapman & Hall, Data set 285 (p. 229)

References

Venables, W. N. and Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth edition. Springer.