granova: Graphical Analysis of Variance

This small collection of functions provides what we call elemental graphics for display of analysis of variance results, David C. Hoaglin, Frederick Mosteller and John W. Tukey (1991, ISBN:978-0-471-52735-0), Paul R. Rosenbaum (1989) doi:10.2307/2684513 , Robert M. Pruzek and James E. Helmreich https://jse.amstat.org/v17n1/helmreich.html. The term elemental derives from the fact that each function is aimed at construction of graphical displays that afford direct visualizations of data with respect to the fundamental questions that drive the particular analysis of variance methods. These functions can be particularly helpful for students and non-statistician analysts. But these methods should be quite generally helpful for work-a-day applications of all kinds, as they can help to identify outliers, clusters or patterns, as well as highlight the role of non-linear transformations of data.

References

David C. Hoaglin, Frederick Mosteller and John W. Tukey (1991, ISBN:978-0-471-52735-0).

Paul R. Rosenbaum (1989) <doi:10.2307/2684513>.

Robert M. Pruzek and James E. Helmreich <https://jse.amstat.org/v17n1/helmreich.html>.

See also

Author

This package has been written by Robert M. Pruzek and James E. Helmreich. Frederic Bertrand is now the maintainer of package and has been doing several code fixes and improvements to the package.

Maintainer: Frederic Bertrand <frederic.bertrand@lecnam.net>