Poison data from Biological Experiment

Survial times of animals in a 3 x 4 factorial experiment involving poisons (3 levels) and various treatments (four levels), as described in Chapter 8 of Box, Hunter and Hunter.

Usage

data(poison)

Format

This data frame was originally poison.data from the package BHH2, but as presented here has added columns; no NAs.

Poison

Factor with three levels I, II, and III.

Treatment

Factor with four levels, A, B, C, and D.

Group

Factor with 12 levels, 1:12.

SurvTime

Numeric; survival time.

RateSurvTime

Numeric; inverse of SurvTime

RankRateSurvTime

Numeric; RateSurvTime scores have been converted to ranks, and then rescaled to have the same median as and a spread comparable to RateSurvTime

Source

Box, G. E. P. and D. R. Cox, An Analysis of Transformations (with discussion), Journal of the Royal Statistical Society, Series B, Vol. 26, No. 2, pp. 211 - 254.

References

Box G. E. P, Hunter, J. S. and Hunter, W. C. (2005). Statistics for Experimenters II. New York: Wiley.